Initial implementation of client methods
The Fortran SILC client relies on interfaces to the C-client. However
for ease of development and for improved user experience, the Fortran
client more closely mimics the C++ client by taking advantage of
object-oriented features of Fortran2003. The iso_c_binding module is
also used to facilitate interoperability with C.

Update fortran enum for unsupported options

Some variable types are not supported in Fortran. We want to include
these in enums in the Fortran list to help keep the clients
synchronized, but these are now set to -1 to represent an invalid
choice.