Implement Fortran client script and model methods
The SILC API used to interact with the scripts and models have been
implemented in the Fortran client. The major Fortran-C interop issues
that this addresses is passing of 2D character arrays. First, the arrays
are copied to an equivalently sized, c_char type array. Then a c_ptr is
created that points to that array. Lastly, this c_ptr is passed 'by
value' (i.e. dereferenced) so that on the c-side, it is the pointer to
an equivalent **char array.