Return pointer in get_meta_scalars in C
Due to limitations in the Fortran-C interop, memory that has
been allocated by a C-function, must be returned by the
function and cannot be returned as an argument. Essentially,
this is because Fortran has not allocated the memory, C will
allocate random memory somewhere, and the address that's
returned is somewhere unknown in memory.

Instead we now return a void* pointer which is then
transferred from a C-type pointer to Fortran pointer.

Implement dataset methods

Most of the functionality associated with the Dataset type
has been implemented in the Fortran client. This includes,
setting/unpacking/getting a tensor and scalar metadata.

Refactor type declarations

The Fortran SILC client and dataset modules have been
renamed to silc_client and silc_dataset with the derived
types called client_type and dataset_type.

Restructure Fortran client source code

The source code for the Fortran client is now laid out into
subdirectories for the client and dataset implementations.
This is primarily done to avoid the proliferation of
similarly named files and to make it easier to read and
maintain the primary client.F90 and dataset.F90 source code