Data aggregation C++ API (#248)
An initial API for dataset aggregation has been implemented 
in the C++ client layer. As part of the dataset aggregation 
capabilities, the client now has the internal ability (not user facing)
to execute Redis pipelines.  This aggregation API will be extended
to Fortran, C and Python clients in a future PR.  Additionally, a
multi-threaded version of the internal pipeline execution in a future
PR will significantly increase aggregation performance.

[ committed by @mellis13 ]
[ reviewed by @ashao @billschereriii  ]