Implement multi-GPU C/Fortran methods (#254)
The C and Fortran clients now have access to all of the multigpu
helper functions implemented in the C++ clients. This notably 
includes `set_model_multigpu` and `run_model_gpu` which
allows for the storage and execution of ML models for inference
on systesm with multiple GPU devices. This PR also includes a 
variant of the MNIST test that uses most of these methods.

[ commited by @ashao ]
[ reviewed by @billschereriii, @mellis13 ]