Implement dataset aggregation for Fortran (#258)
Adds in all methods and C interfaces to enable the dataset
aggregation functionality in Fortran. Most of the transformations
and implementations are similar to those done before. The one
departure from the C implementation is that for returning a list
of datasets, we rely on Fortran to allocate the necessary memory
for the list of dataset objects. This vector is then populated
within the C client without allocating a new vector.

The Fortran tests for this functionality creates 5 new datasets and
adds a tensor to them filled with random numbers. These are then
individually added to the database and then appended to the
dataset list. The list is then retrieved from the database and the
tensors are checked to make sure that they are identical

[ committed by @ashao ]
[ reviewed by @billschereriii @mellis13 ]